Move server to server receipt info definition closer to where it's used.
Add links to the documentation as it's getting hard to keep track of what's what.
