+++ typescript/src/pubsub/apple.ts
@@ -1,12 +1,34 @@
 import 'source-map-support/register'
-import {ONE_YEAR_IN_SECONDS, parseStoreAndSend} from "./pubsub";
+import {parseStoreAndSend} from "./pubsub";
 import {SubscriptionEvent} from "../models/subscriptionEvent";
-import {AppleReceiptInfo} from "../models/appleReceiptInfo";
 import {dateToSecondTimestamp, thirtyMonths} from "../utils/dates";
 import {AppleSubscriptionReference} from "../models/subscriptionReference";
 import {APIGatewayProxyEvent, APIGatewayProxyResult} from "aws-lambda";
 import {Option} from "../utils/option";
 import {fromAppleBundle} from "../services/appToPlatform";
+
+// this is the definition of a receipt as received by the server to server notification system.
+// Not to be confused with apple's receipt validation receipt info (although they do look similar, they are different)
+// See https://developer.apple.com/documentation/appstoreservernotifications/responsebody
+export interface AppleReceiptInfo {
+    transaction_id: string,
+    bid: string,
+    product_id: string,
+    original_transaction_id: string,
+    item_id: string,
+    app_item_id: string,
+    web_order_line_item_id: string,
+    unique_identifier: string,
+    unique_vendor_identifier: string,
+    quantity: string,
+    purchase_date_ms: string,
+    original_purchase_date_ms: string,
+    expires_date: string,
+    is_in_intro_offer_period: string,
+    is_trial_period: string,
+    bvrs: string,
+    version_external_identifier: string
+}
 
 export interface StatusUpdateNotification {
     environment: string,
