+++ typescript/src/services/appleValidateReceipts.ts
@@ -34,6 +34,9 @@ export interface ValidationOptions {
 
 // there are more fields, I cherry picked what was relevant
 // https://developer.apple.com/library/archive/releasenotes/General/ValidateAppStoreReceipt/Chapters/ValidateRemotely.html
+// And
+// https://developer.apple.com/documentation/appstoreservernotifications/responsebody
+// Both documentation are equally wrong when compared to the reality, so some of the definition bellow match neither.
 export interface AppleValidationServerResponse {
     auto_renew_status: 0 | 1,
     "is-retryable"?: boolean,
